Add security condition to deleteImage function to avoid others users can delete images not belonging to them

+++ src/app/components/images/images.component.ts
@@ -70,9 +70,12 @@ export class ImagesComponent implements OnInit {
   }
 
   deleteImage = (image: Image, e: Event) => {
-    this.fbs.deleteImage(image).then(() => {
-      this.openSnackBar(`Image '${image.name}' deleted `,'Ok')
-    });   
+    // Security condition    
+    if (this.user.id === image.user_id){
+      this.fbs.deleteImage(image).then(() => {
+        this.openSnackBar(`Image '${image.name}' deleted `,'Ok')
+      });   
+    }    
     // Avoid open image modal window at once
     e.stopPropagation();
   }
